requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
spacy==3.7.2
python-dateutil==2.8.2
reportlab==4.0.7
openpyxl==3.1.2

# spaCy English model (so Streamlit Cloud doesn't have to download it)
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl